use Track instead of LastFMTrack in album, change seconds func to return Int instead of Double

+++ LastFM/Models/Artist.swift
@@ -61,8 +61,8 @@ class Artist: Decodable, CustomStringConvertible {
         self.cover = .url(url!)
     }
     
-    static func seconds(_ min: Int = 0, _ sec: Int) -> Double {
-        return Double(min * 60 + sec)
+    static func seconds(_ min: Int = 0, _ sec: Int) -> Int {
+        return min * 60 + sec
     }
 }
